+++ Assert.Domain/Entities/TState.cs
@@ -5,9 +5,9 @@ namespace Assert.Domain.Entities;
 
 public partial class TState
 {
-    public long StateId { get; set; }
+    public int StateId { get; set; }
 
-    public long CountryId { get; set; }
+    public int CountryId { get; set; }
 
     public string? IataCode { get; set; }
