+++ Assert.Domain/Entities/TpPropertyAddress.cs
@@ -11,7 +11,7 @@ public partial class TpPropertyAddress
 
     public string? ZipCode { get; set; }
 
-    public long? CityId { get; set; }
+    public int? CityId { get; set; }
 
     public string? Address1 { get; set; }
