+++ Assert.Domain/Models/ListingProcessDataModel.cs
@@ -48,7 +48,7 @@
         public string? Address1 { get; set; }
 
         public string? Address2 { get; set; }
-        public long? CityId { get; set; }
+        public int? CityId { get; set; }
     }
     public class ProcessData_SpaceModel
     {
